move RouterType to types package
